umdostestes

+++ backend/src/main/java/com/votzz/backend/dto/AdminDashboardStats.java
@@ -1,10 +1,12 @@
 package com.votzz.backend.dto;
 
+import java.math.BigDecimal;
+
 public record AdminDashboardStats(
     long totalUsers,
     long onlineUsers,
     long totalTenants,
     long activeTenants,
-    double mrr,
+    BigDecimal mrr, // Mudado para BigDecimal para precisão financeira
     long latencyMs 
 ) {}
